fix: check if allowServices is false and set it to true
We need to do that if we want to create a conversation with bot. We need to set the allowServices to true. Then we're going to add service to the accessRoles and the payload would be correct. The reason wasn't working was because we didn't add "service" to the accessRoles so the payload was incorrect.